requirements.txt
playwright
python-dotenv
coloredlogs
tk
